Refactor Step methods into separate interfaces
